Update object_track.py
I did this on linux and i narrowed it down to where it can track the ball when it's pure white and when it gets a bit darker. i almost got rid of the blobs on the other stuff other than the baseball. after running the code two windows will pop up. One for regular tracking and another to show you when the ball is pure white. to close or stop program i just did ctrl +c on linux.